+++ tests/ops/service/privacy_request/request_runner_service_test.py
@@ -6,6 +6,7 @@ from uuid import uuid4
 
 import pydash
 import pytest
+from fideslib.models.audit_log import AuditLog, AuditLogAction
 from pydantic import ValidationError
 from sqlalchemy import column, select, table
 from sqlalchemy.orm import Session
@@ -288,6 +289,17 @@ def test_create_and_process_access_request(
     assert results[visit_key][0]["email"] == customer_email
     log_id = pr.execution_logs[0].id
     pr_id = pr.id
+
+    finished_audit_log: AuditLog = AuditLog.filter(
+        db=db,
+        conditions=(
+            (AuditLog.privacy_request_id == pr_id)
+            & (AuditLog.action == AuditLogAction.finished)
+        ),
+    ).first()
+
+    assert finished_audit_log is not None
+
     # Both pre-execution webhooks and both post-execution webhooks were called
     assert trigger_webhook_mock.call_count == 4